recommender_system_project/requirements.txt
Flask
pandas
scikit-surprise
